[Bug #7910] - MPR and Fusion do not account for varied Modality LUTs in source images [Bug #9266] - Unable to render images with very small rescale slope (<<1)  * There do not seem to be any use cases for polymorphic LUTs that can be combined in an arbitrary manner for the purposes of the image display pipeline. The standard grayscale display pipeline consists of, in sequence, a modality LUT, a renormalization function, a VOI LUT, and a color map, with well defined interfaces between successive LUTs that need not be the same. In generic ILut interface allowed these functions to be combined in a different order although, in practice, a number of restrictions built into the implementations (which already fall into separate families of modality LUTs, VOI LUTs and color maps) prevented them from being combined in arbitrary manners (e.g. color maps throw an exception if a LUT is installed after it). As a result, a number of refactorings are being undertaken to restructure the types used in the display pipeline.   * IComposableLut is now a lookup of floating-point outputs from floating-point inputs.    * IModalityLut is a specialization of IComposableLut to perform a lookup of floating-point outputs from integer inputs.    * IVoiLut is a specialization of IComposableLut to perform a lookup of integer outputs from floating-point inputs.    * Various base classes for IModalityLut and IVoiLut have been converted to allow minimal changes in derived classes.

+++ ImageViewer/Tools/Standard/ProbeTool.cs
@@ -174,7 +174,7 @@ namespace ClearCanvas.ImageViewer.Tools.Standard
 						GrayscaleImageGraphic image = _selectedImageGraphic as GrayscaleImageGraphic;
 
 						int pixelValue = 0;
-						int modalityLutValue = 0;
+						double modalityLutValue = 0;
 						int voiLutValue = 0;
 
 						GetPixelValue(image, sourcePointRounded, ref pixelValue, ref pixelValueString);
@@ -228,7 +228,7 @@ namespace ClearCanvas.ImageViewer.Tools.Standard
 		private void GetModalityLutValue(
 			GrayscaleImageGraphic grayscaleImage,
 			int pixelValue,
-			ref int modalityLutValue,
+			ref double modalityLutValue,
 			ref string modalityLutString)
 		{
 			if (grayscaleImage.ModalityLut != null)
@@ -245,7 +245,7 @@ namespace ClearCanvas.ImageViewer.Tools.Standard
 
 		private void GetVoiLutValue(
 			GrayscaleImageGraphic grayscaleImage,
-			int modalityLutValue,
+			double modalityLutValue,
 			ref int voiLutValue,
 			ref string voiLutString)
 		{
